\name{MC2toPath-package}
\alias{MC2toPath-package}
\alias{MC2toPath}
\docType{package}
\title{
Translate MC2 output netcdf files into inter-PVT transition rates.
}
\description{
Translate MC2 output netcdf files into inter-PVT transition rates.
}
\details{
\tabular{ll}{
Package: \tab MC2toPath\cr
Type: \tab Package\cr
Version: \tab 0.0.16
\cr
Date: \tab 2014-05-16\cr
License: \tab GPL-2\cr
}
Post processes MC2 output, especially for use by Path or ST-Sim.  A detailed example is given in the document, "MakingThePATHmegamodel.pdf" in inst/doc.
}
\author{

Author: Dave Conklin and Emilie Henderson
Maintainer: Dave Conklin<david.conklin@mac.com>
}

\keyword{ package }

