\name{vegChanges_step11}
\docType{data}
\alias{vegChanges_step11}
\title{Example data for step 11 in the MC2toPath package}
\description{
The list object vegChanges_step11 is as produced 
by the call to VegTypeChanges() in step 11 of the example script in VegTypeChanges.Rd.  
The actual call to VegTypeChanges() in the example script has been commented out 
to save execution time in CRAN.
}
\usage{data(vegChanges_step11)}
\format{ a list object of length 6, containing:
 tgtFile, years, vts2keep, vtFracsReduced, changeFracsReduced, and vt2pvtlut
}
\source{Example step 11 run on WW2100_HadGEM2-ES85_year_sample.nc}
\references{doc/MakingThePATHmegamodel.pdf}
\keyword{datasets}
